+++ src/pdfjs-worker.js
@@ -1,0 +1,7 @@
+import { pdfjs } from 'react-pdf';
+
+// Set the worker source
+pdfjs.GlobalWorkerOptions.workerSrc = new URL(
+  'pdfjs-dist/build/pdf.worker.min.js',
+  import.meta.url,
+).toString();
